backend/requirements.txt
Django==4.1.7
djangorestframework
django-environ==0.7.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.3
daphne==4.0.0
channels==4.0.0
djangorestframework-simplejwt
requests
httpx
djoser
drf-nested-routers
Pillow
gunicorn
dj-database-url
whitenoise
